+++ trixnity-messenger/src/androidMain/kotlin/de/connect2x/trixnity/messenger/util/Images.android.kt
@@ -1,9 +1,9 @@
 package de.connect2x.trixnity.messenger.util
 
 import android.graphics.BitmapFactory
-import de.connect2x.trixnity.messenger.viewmodel.util.limitedByteArrayOrNull
 import io.ktor.http.*
 import net.folivo.trixnity.utils.ByteArrayFlow
+import net.folivo.trixnity.utils.toByteArray
 import org.koin.core.module.Module
 import org.koin.dsl.module
 
@@ -21,7 +21,7 @@ suspend fun getImageDimensions(
     maxMediaSize: Long,
     mimeType: ContentType?
 ): Pair<Int?, Int?> {
-    val byteArray = byteArrayFlow.limitedByteArrayOrNull(maxMediaSize)
+    val byteArray = byteArrayFlow.toByteArray(maxMediaSize)
     val bitmap = byteArray?.let { BitmapFactory.decodeByteArray(it, 0, it.size) }
     return bitmap?.width to bitmap?.height
 }