+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/viewmodel/roomlist/AccountViewModel.kt
@@ -70,9 +70,9 @@ open class AccountViewModelImpl(
     private val initials = get<Initials>()
     private val messengerSettings = get<MatrixMessengerSettingsHolder>()
 
-    private val maxAvatarSize = get<MatrixMessengerConfiguration>().avatarMaxSize
+    private val maxMediaSizeInMemory = get<MatrixMessengerConfiguration>().maxMediaSizeInMemory
     override val accounts: StateFlow<List<AccountInfo>> =
-        matrixClients.toAccountInfo(messengerSettings, initials, maxAvatarSize)
+        matrixClients.toAccountInfo(coroutineScope, messengerSettings, initials, maxMediaSizeInMemory)
             .stateIn(coroutineScope, SharingStarted.WhileSubscribed(), listOf())
 
     override val activeAccount: StateFlow<UserId?> =